WIP Mint Asset Circuit, also includes public address added to asset info
