+++ ironfish-rust/src/proofs/circuit/mod.rs
@@ -1,3 +1,6 @@
 pub mod create_asset;
+pub mod mint_asset;
 pub mod mycircuit;
+pub mod output;
 pub mod sapling;
+pub mod spend;
