+++ ironfish-rust/src/proofs/circuit/spend.rs
@@ -1,0 +1,358 @@
+use ff::PrimeField;
+use group::Curve;
+
+use bellman::{Circuit, ConstraintSystem, SynthesisError};
+
+use jubjub::ExtendedPoint;
+use zcash_primitives::constants;
+
+use zcash_primitives::primitives::{PaymentAddress, ProofGenerationKey};
+
+use bellman::gadgets::blake2s;
+use bellman::gadgets::boolean::{self, AllocatedBit, Boolean};
+use bellman::gadgets::multipack;
+use bellman::gadgets::num;
+use bellman::gadgets::Assignment;
+use zcash_proofs::circuit::ecc::EdwardsPoint;
+use zcash_proofs::circuit::{ecc, pedersen_hash};
+use zcash_proofs::constants::{
+    NOTE_COMMITMENT_RANDOMNESS_GENERATOR, NULLIFIER_POSITION_GENERATOR,
+    PROOF_GENERATION_KEY_GENERATOR, SPENDING_KEY_GENERATOR, VALUE_COMMITMENT_RANDOMNESS_GENERATOR,
+};
+
+use crate::primitives::asset_type::AssetType;
+use crate::primitives::sapling::ValueCommitment;
+use crate::proofs::circuit::sapling::expose_value_commitment;
+
+pub struct Spend {
+    // TODO: Should we pass this in anymore, or just rely on asset type? Feels like this could lead to accidental bugs.
+    /// Pedersen commitment to the value being spent
+    pub value_commitment: Option<ValueCommitment>,
+
+    /// Asset type that the value is denominated in
+    pub asset_type: Option<AssetType>,
+
+    /// Key required to construct proofs for spending notes
+    /// for a particular spending key
+    pub proof_generation_key: Option<ProofGenerationKey>,
+
+    /// The payment address associated with the note
+    pub payment_address: Option<PaymentAddress>,
+
+    /// The randomness of the note commitment
+    pub commitment_randomness: Option<jubjub::Fr>,
+
+    /// Re-randomization of the public key
+    pub ar: Option<jubjub::Fr>,
+
+    /// The authentication path of the commitment in the tree
+    pub auth_path: Vec<Option<(bls12_381::Scalar, bool)>>,
+
+    /// The anchor; the root of the tree. If the note being
+    /// spent is zero-value, this can be anything.
+    pub anchor: Option<bls12_381::Scalar>,
+}
+impl Circuit<bls12_381::Scalar> for Spend {
+    fn synthesize<CS: ConstraintSystem<bls12_381::Scalar>>(
+        self,
+        cs: &mut CS,
+    ) -> Result<(), SynthesisError> {
+        // Prover witnesses ak (ensures that it's on the curve)
+        let ak = ecc::EdwardsPoint::witness(
+            cs.namespace(|| "ak"),
+            self.proof_generation_key.as_ref().map(|k| k.ak.into()),
+        )?;
+
+        // There are no sensible attacks on small order points
+        // of ak (that we're aware of!) but it's a cheap check,
+        // so we do it.
+        ak.assert_not_small_order(cs.namespace(|| "ak not small order"))?;
+
+        // Rerandomize ak and expose it as an input to the circuit
+        {
+            let ar = boolean::field_into_boolean_vec_le(cs.namespace(|| "ar"), self.ar)?;
+
+            // Compute the randomness in the exponent
+            let ar = ecc::fixed_base_multiplication(
+                cs.namespace(|| "computation of randomization for the signing key"),
+                &SPENDING_KEY_GENERATOR,
+                &ar,
+            )?;
+
+            let rk = ak.add(cs.namespace(|| "computation of rk"), &ar)?;
+
+            rk.inputize(cs.namespace(|| "rk"))?;
+        }
+
+        // Compute nk = [nsk] ProofGenerationKey
+        let nk;
+        {
+            // Witness nsk as bits
+            let nsk = boolean::field_into_boolean_vec_le(
+                cs.namespace(|| "nsk"),
+                self.proof_generation_key.as_ref().map(|k| k.nsk),
+            )?;
+
+            // NB: We don't ensure that the bit representation of nsk
+            // is "in the field" (jubjub::Fr) because it's not used
+            // except to demonstrate the prover knows it. If they know
+            // a congruency then that's equivalent.
+
+            // Compute nk = [nsk] ProvingPublicKey
+            nk = ecc::fixed_base_multiplication(
+                cs.namespace(|| "computation of nk"),
+                &PROOF_GENERATION_KEY_GENERATOR,
+                &nsk,
+            )?;
+        }
+
+        // This is the "viewing key" preimage for CRH^ivk
+        let mut ivk_preimage = vec![];
+
+        // Place ak in the preimage for CRH^ivk
+        ivk_preimage.extend(ak.repr(cs.namespace(|| "representation of ak"))?);
+
+        // This is the nullifier preimage for PRF^nf
+        let mut nf_preimage = vec![];
+
+        // Extend ivk and nf preimages with the representation of
+        // nk.
+        {
+            let repr_nk = nk.repr(cs.namespace(|| "representation of nk"))?;
+
+            ivk_preimage.extend(repr_nk.iter().cloned());
+            nf_preimage.extend(repr_nk);
+        }
+
+        assert_eq!(ivk_preimage.len(), 512);
+        assert_eq!(nf_preimage.len(), 256);
+
+        // Compute the incoming viewing key ivk
+        let mut ivk = blake2s::blake2s(
+            cs.namespace(|| "computation of ivk"),
+            &ivk_preimage,
+            constants::CRH_IVK_PERSONALIZATION,
+        )?;
+
+        // drop_5 to ensure it's in the field
+        ivk.truncate(jubjub::Fr::CAPACITY as usize);
+
+        // Witness g_d, checking that it's on the curve.
+        let g_d = {
+            ecc::EdwardsPoint::witness(
+                cs.namespace(|| "witness g_d"),
+                self.payment_address
+                    .as_ref()
+                    .and_then(|a| a.g_d().map(jubjub::ExtendedPoint::from)),
+            )?
+        };
+
+        // Check that g_d is not small order. Technically, this check
+        // is already done in the Output circuit, and this proof ensures
+        // g_d is bound to a product of that check, but for defense in
+        // depth let's check it anyway. It's cheap.
+        g_d.assert_not_small_order(cs.namespace(|| "g_d not small order"))?;
+
+        // Compute pk_d = g_d^ivk
+        let pk_d = g_d.mul(cs.namespace(|| "compute pk_d"), &ivk)?;
+
+        // Witness the asset type
+        // TODO: Does this properly verify that the spend note is the right asset generator?
+        // Could this be spoofed, or does this not matter? Need to consider what's public/private
+        // In other words: Does this verify that the actual note's asset generator is valid
+        // Or are we allowing this to be _any_ generator
+        let asset_generator = ecc::EdwardsPoint::witness(
+            cs.namespace(|| "asset_generator"),
+            self.asset_type
+                .as_ref()
+                .and_then(|at| at.asset_generator().into()),
+        )?;
+
+        let value_commitment_generator = ecc::EdwardsPoint::witness(
+            cs.namespace(|| "value commitment generator"),
+            self.asset_type
+                .as_ref()
+                .and_then(|at| ExtendedPoint::from(at.value_commitment_generator()).into()),
+        )?;
+
+        value_commitment_generator.assert_not_small_order(
+            cs.namespace(|| "value_commitment_generator not small order"),
+        )?;
+
+        // Compute note contents:
+        // asset_generator, value (in big endian), g_d, pk_d
+        let mut note_contents = vec![];
+
+        // Place asset_generator in the note
+        note_contents
+            .extend(asset_generator.repr(cs.namespace(|| "representation of asset_generator"))?);
+
+        // Handle the value; we'll need it later for the
+        // dummy input check.
+        let mut value_num = num::Num::zero();
+        {
+            // Get the value in little-endian bit order
+            let value_bits = expose_value_commitment(
+                cs.namespace(|| "value commitment"),
+                value_commitment_generator,
+                self.value_commitment,
+            )?;
+
+            // Compute the note's value as a linear combination
+            // of the bits.
+            let mut coeff = bls12_381::Scalar::one();
+            for bit in &value_bits {
+                value_num = value_num.add_bool_with_coeff(CS::one(), bit, coeff);
+                coeff = coeff.double();
+            }
+
+            // Place the value in the note
+            note_contents.extend(value_bits);
+        }
+
+        // Place g_d in the note
+        note_contents.extend(g_d.repr(cs.namespace(|| "representation of g_d"))?);
+
+        // Place pk_d in the note
+        note_contents.extend(pk_d.repr(cs.namespace(|| "representation of pk_d"))?);
+
+        assert_eq!(
+            note_contents.len(),
+            256 + // asset_generator
+            64 + // value
+            256 + // g_d
+            256 // pk_d
+        );
+
+        // Compute the hash of the note contents
+        let mut cm = pedersen_hash::pedersen_hash(
+            cs.namespace(|| "note content hash"),
+            pedersen_hash::Personalization::NoteCommitment,
+            &note_contents,
+        )?;
+
+        {
+            // Booleanize the randomness for the note commitment
+            let rcm = boolean::field_into_boolean_vec_le(
+                cs.namespace(|| "rcm"),
+                self.commitment_randomness,
+            )?;
+
+            // Compute the note commitment randomness in the exponent
+            let rcm = ecc::fixed_base_multiplication(
+                cs.namespace(|| "computation of commitment randomness"),
+                &NOTE_COMMITMENT_RANDOMNESS_GENERATOR,
+                &rcm,
+            )?;
+
+            // Randomize the note commitment. Pedersen hashes are not
+            // themselves hiding commitments.
+            cm = cm.add(cs.namespace(|| "randomization of note commitment"), &rcm)?;
+        }
+
+        // This will store (least significant bit first)
+        // the position of the note in the tree, for use
+        // in nullifier computation.
+        let mut position_bits = vec![];
+
+        // This is an injective encoding, as cur is a
+        // point in the prime order subgroup.
+        let mut cur = cm.get_u().clone();
+
+        // Ascend the merkle tree authentication path
+        for (i, e) in self.auth_path.into_iter().enumerate() {
+            let cs = &mut cs.namespace(|| format!("merkle tree hash {}", i));
+
+            // Determines if the current subtree is the "right" leaf at this
+            // depth of the tree.
+            let cur_is_right = boolean::Boolean::from(boolean::AllocatedBit::alloc(
+                cs.namespace(|| "position bit"),
+                e.map(|e| e.1),
+            )?);
+
+            // Push this boolean for nullifier computation later
+            position_bits.push(cur_is_right.clone());
+
+            // Witness the authentication path element adjacent
+            // at this depth.
+            let path_element =
+                num::AllocatedNum::alloc(cs.namespace(|| "path element"), || Ok(e.get()?.0))?;
+
+            // Swap the two if the current subtree is on the right
+            let (ul, ur) = num::AllocatedNum::conditionally_reverse(
+                cs.namespace(|| "conditional reversal of preimage"),
+                &cur,
+                &path_element,
+                &cur_is_right,
+            )?;
+
+            // We don't need to be strict, because the function is
+            // collision-resistant. If the prover witnesses a congruency,
+            // they will be unable to find an authentication path in the
+            // tree with high probability.
+            let mut preimage = vec![];
+            preimage.extend(ul.to_bits_le(cs.namespace(|| "ul into bits"))?);
+            preimage.extend(ur.to_bits_le(cs.namespace(|| "ur into bits"))?);
+
+            // Compute the new subtree value
+            cur = pedersen_hash::pedersen_hash(
+                cs.namespace(|| "computation of pedersen hash"),
+                pedersen_hash::Personalization::MerkleTree(i),
+                &preimage,
+            )?
+            .get_u()
+            .clone(); // Injective encoding
+        }
+
+        {
+            let real_anchor_value = self.anchor;
+
+            // Allocate the "real" anchor that will be exposed.
+            let rt = num::AllocatedNum::alloc(cs.namespace(|| "conditional anchor"), || {
+                Ok(*real_anchor_value.get()?)
+            })?;
+
+            // (cur - rt) * value = 0
+            // if value is zero, cur and rt can be different
+            // if value is nonzero, they must be equal
+            cs.enforce(
+                || "conditionally enforce correct root",
+                |lc| lc + cur.get_variable() - rt.get_variable(),
+                |lc| lc + &value_num.lc(bls12_381::Scalar::one()),
+                |lc| lc,
+            );
+
+            // Expose the anchor
+            rt.inputize(cs.namespace(|| "anchor"))?;
+        }
+
+        // Compute the cm + g^position for preventing
+        // faerie gold attacks
+        let mut rho = cm;
+        {
+            // Compute the position in the exponent
+            let position = ecc::fixed_base_multiplication(
+                cs.namespace(|| "g^position"),
+                &NULLIFIER_POSITION_GENERATOR,
+                &position_bits,
+            )?;
+
+            // Add the position to the commitment
+            rho = rho.add(cs.namespace(|| "faerie gold prevention"), &position)?;
+        }
+
+        // Let's compute nf = BLAKE2s(nk || rho)
+        nf_preimage.extend(rho.repr(cs.namespace(|| "representation of rho"))?);
+
+        assert_eq!(nf_preimage.len(), 512);
+
+        // Compute nf
+        let nf = blake2s::blake2s(
+            cs.namespace(|| "nf computation"),
+            &nf_preimage,
+            constants::PRF_NF_PERSONALIZATION,
+        )?;
+
+        multipack::pack_into_inputs(cs.namespace(|| "pack nullifier"), &nf)
+    }
+}
